backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
langchain
langchain-openai
langchain-community
openai
python-dotenv
